gen-val/src/generation/src/NIST.CVP.ACVTS.Libraries.Generation/SHA3/v1_0/TestGroupGeneratorMonteCarlo.cs
using System.Collections.Generic;
using System.Threading.Tasks;
using NIST.CVP.ACVTS.Libraries.Crypto.Common.Hash.ShaWrapper.Helpers;
using NIST.CVP.ACVTS.Libraries.Generation.Core;
using NIST.CVP.ACVTS.Libraries.Math.Domain;

namespace NIST.CVP.ACVTS.Libraries.Generation.SHA3.v1_0
{
    public class TestGroupGeneratorMonteCarlo : ITestGroupGeneratorAsync<Parameters, TestGroup, TestCase>
    {
        public const string TEST_TYPE = "MCT";

        public Task<List<TestGroup>> BuildTestGroupsAsync(Parameters parameters)
        {
            var testGroups = new List<TestGroup>();
            var function = ShaAttributes.GetHashFunctionFromName(parameters.Algorithm);

            foreach (var digestSize in parameters.DigestSizes)        // Only ever as single digest size
            {
                // Must make sure we actually have a domain. SHAKE will, SHA3 will not.
                MathDomain domain = null;
                if (parameters.OutputLength != null)
                {
                    domain = parameters.OutputLength.GetDeepCopy();
                }

                var testGroup = new TestGroup
                {
                    Function = function.Mode,
                    DigestSize = function.DigestSize,
                    BitOrientedInput = parameters.BitOrientedInput,
                    BitOrientedOutput = parameters.BitOrientedOutput,
                    OutputLength = domain,
                    TestType = TEST_TYPE
                };

                testGroups.Add(testGroup);
            }

            return Task.FromResult(testGroups);
        }
    }
}
