gen-val/src/generation/test/NIST.CVP.ACVTS.Libraries.Generation.Tests/SHA3/v1_0/TestGroupGeneratorMonteCarloTests.cs
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;
using NIST.CVP.ACVTS.Libraries.Generation.SHA3.v1_0;
using NIST.CVP.ACVTS.Tests.Core.TestCategoryAttributes;
using NUnit.Framework;

namespace NIST.CVP.ACVTS.Libraries.Generation.Tests.SHA3.v1_0
{
    [TestFixture, UnitTest]
    public class TestGroupGeneratorMonteCarloTests
    {
        // These tests don't really mean anything when the only digest size is the one in the algorithm name
        private static object[] parameters =
        {
            new object[]
            {
                0, // 1 * 1
                new ParameterValidatorTests.ParameterBuilder()
                    .WithDigestSizes(new List<int>() { }) // 0
                    .WithAlgorithm("SHA3-256")  // 1
                    .Build()
            }
        };
        [Test]
        [TestCaseSource(nameof(parameters))]
        public async Task ShouldCreate1TestGroupForEachCombinationOfModeAndDigestSize(int expectedGroupsCreated, Parameters parameters)
        {
            var subject = new TestGroupGeneratorMonteCarlo();
            var results = await subject.BuildTestGroupsAsync(parameters);
            Assert.AreEqual(expectedGroupsCreated, results.Count());
        }
    }
}
